First version of the reservation screen
This is the first version of the reservation screen. It was developed reusing part of code from the previous attendance component (previously named reservations as well). There are still some fixes to be done to it.

eg:
 - setStops is called more times than expected
 - The line dots pics must be improved
 - The submit button is missing

Be careful:
Some files were renamed but for some reasons it was recorded instead as deletion and creation of files.

New to the project:
- Component Reservations with the main layout
- Component Stop-lines and its childs used for the line layout
- Service Reservations to fill the reservation component and get
information on the selected stop.
- Service Stop used to create the line and comunicate between the 2
components.